refactor: move EditorConfig units to src/EditorConfig subdir
