+++ src/EditorConfig/EditorConfig.cpp
