+++ src/EditorConfig/EditorConfig.hpp
